requirements.txt
# Environment
python-dotenv

# Server
fastmcp>=2.0.0
fastapi>=0.110.0
mcp[cli]>=1.0.0
uvicorn[standard]>=0.29.0

# Client
langchain
langchain-google-genai
langchain-mcp-adapters